// Frontend/src/app/models/API/Request/compose-request-request.interface.ts
export interface ComposeRequestRequest {
  requestorServer: string;
  requestorName: string;
  clubName: string;
  clubDatacenter: string;
  clubServer: string;
  clubDistrict: string;
  clubWard: number;
  clubPlot: number;
  clubHours: string;
  info: string;
  clubWebsite: string;
  captchaToken: string;
}